+++ base/task.go
@@ -2,10 +2,10 @@ package base
 
 type Task struct {
 	Machine Machine `json:"machine"`
-	Time    int     `json:"time"`
+	Time    uint64     `json:"time"`
 }
 
-func (t Task) Deconstruct() (Machine, int) {
+func (t Task) Deconstruct() (Machine, uint64) {
 	return t.Machine, t.Time
 }
